+++ src/image.h
@@ -1,0 +1,10 @@
+#ifndef _XD_IMAGE_H_
+#define _XD_IMAGE_H_
+
+typedef struct {
+	int width;
+	int height;
+} IMAGE_SIZE;
+
+
+#endif