+++ test/example1.js
@@ -3,7 +3,7 @@ const synchttp = require('../src/build/Release/binding');
 let content = "";
 try{
 	content = synchttp.http_get("http://www.baidu.com",10);
-	console.log(content);
+	//console.log(content);
 	console.log("\nRecv Length:"+content.length);
 }
 catch(err)
@@ -11,4 +11,12 @@ catch(err)
 	console.log(err);
 }
 
-
+let imgsize = new Object();
+try{
+	imgsize = synchttp.get_image_size("http://wmod.66rpg.com/res/loading.gif");
+	console.log(imgsize);
+}
+catch(err)
+{
+	console.log(err);
+}